test(soci): fold sudo command tests into test_convert/test_push
